♻️   F04: ORDERPAGE > Refactoring of components names and positions
