+++ src/components/pages/order/OrderBody.js
@@ -1,16 +1,17 @@
 import React from "react";
 import styled from "styled-components";
-import { theme } from "../../../theme";
+import Main from "./Main";
+import Navbar from "./Navbar";
 
 export default function OrderBody() {
-  return <OrderBodyStyled></OrderBodyStyled>;
+  return (
+    <OrderBodyStyled>
+      <Navbar />
+      <Main />
+    </OrderBodyStyled>
+  );
 }
 
 const OrderBodyStyled = styled.div`
-  background-color: ${theme.colors.background_white};
-  border-bottom-left-radius: ${theme.borderRadius.extraRound};
-  border-bottom-right-radius: ${theme.borderRadius.extraRound};
   width: 1400px;
-  height: 85vh;
-  box-shadow: 0px 8px 20px 8px rgba(0, 0, 0, 0.2) inset;
 `;
