+++ src/components/pages/order/RightSide.js
@@ -1,0 +1,13 @@
+import React from "react";
+import styled from "styled-components";
+import UserPanel from "./UserPanel";
+
+export default function RightSide() {
+  return (
+    <RightSideStyled>
+      <UserPanel />
+    </RightSideStyled>
+  );
+}
+
+const RightSideStyled = styled.div``;
